Data Engineering Coursework - Wiki Analyze
This is another sample of programming done through working in an online data engineering course. This project focused on a combination of web scraping and getting a sense for the common words in a wikipedia article. This also allowed me to experiment with what limitations one needs to be aware of when working with larger amounts of information.